msk/acars: add support for noise floor estimator
This introduces a basic, low cpu-cost NF estimator which works by using
an exponential moving average over the magnitude of the received signal
updated only outside of ACARS message processing.

The computational overhead of this extra estimator is virtually nil, as
the signal power estimator has been reworked to use the same input as
the noise floor estimator.

Fixes: https://github.com/TLeconte/acarsdec/issues/84